Update atomWithHash replaceState function to use window.history.state instead of null (#24)
My project uses the new NextJS app router, and we were using atomWithHash to keep track of state. This was breaking when used with the app router when using the back button in the browser. In my project I am just setting the setHash function to do what I committed here. I figured it would be better and easier for anyone else using this library to just set the flag replaceState if they are using the nextjs app router. (Maybe the website/documentation should be updated as well.)

Co-authored-by: Daishi Kato <dai-shi@users.noreply.github.com>